allow for multiple profiles selection in /profile unlink
- move select_answer and select_profile to their matching Cog
- fix type hints

+++ cogs/profile.py
@@ -12,12 +12,12 @@ from discord import app_commands
 from matplotlib import pyplot
 from discord.ext import commands
 
-from classes.ui import ModalProfileLink, ModalProfileUpdate, select_profile
+from classes.ui import ModalProfileLink, SelectProfileView, ModalProfileUpdate, SelectProfilesView
 from utils.funcs import chunker, hero_autocomplete, get_platform_emoji
-from utils.checks import is_premium, has_profile
+from utils.checks import is_premium, has_profile, can_add_profile
 from classes.profile import Profile
 from classes.nickname import Nickname
-from classes.exceptions import CannotCreateGraph
+from classes.exceptions import NoChoice, CannotCreateGraph
 
 if TYPE_CHECKING:
     from asyncpg import Record
@@ -31,10 +31,8 @@ class ProfileCog(commands.Cog, name="Profile"):
 
     profile = app_commands.Group(name="profile", description="Your Overwatch profiles.")
 
-    async def get_profiles(
-        self, interaction: discord.Interaction, member: discord.Member
-    ) -> list[Record]:
-        limit = self.bot.get_profiles_limit(interaction, member.id)
+    async def get_profiles(self, interaction: discord.Interaction, member_id: int) -> list[Record]:
+        limit = self.bot.get_profiles_limit(interaction, member_id)
         query = """SELECT profile.id, platform, username
                    FROM profile
                    INNER JOIN member
@@ -42,7 +40,7 @@ class ProfileCog(commands.Cog, name="Profile"):
                    WHERE member.id = $1
                    LIMIT $2;
                 """
-        return await self.bot.pool.fetch(query, member.id, limit)
+        return await self.bot.pool.fetch(query, member_id, limit)
 
     async def get_profile(self, profile_id: str) -> Record:
         query = """SELECT id, platform, username
@@ -50,6 +48,31 @@ class ProfileCog(commands.Cog, name="Profile"):
                    WHERE id = $1;
                 """
         return await self.bot.pool.fetchrow(query, int(profile_id))
+
+    async def select_profile(
+        interaction: discord.Interaction, message: str, member: None | discord.Member = None
+    ) -> str:
+        member = member or interaction.user
+        profiles = await interaction.client.get_cog("Profile").get_profiles(interaction, member.id)
+
+        # if there only is a profile then just return it
+        if len(profiles) == 1:
+            profile_id, _, _ = profiles[0]
+            return await interaction.client.get_cog("Profile").get_profile(profile_id)
+
+        view = SelectProfileView(profiles, author_id=interaction.user.id)
+
+        if interaction.response.is_done():
+            view.message = await interaction.followup.send(message, view=view)
+        else:
+            view.message = await interaction.response.send_message(message, view=view)
+        await view.wait()
+
+        choice = view.select.values[0] if len(view.select.values) else None
+
+        if choice is not None:
+            return await interaction.client.get_cog("Profile").get_profile(choice)
+        raise NoChoice() from None
 
     async def list_profiles(
         self, interaction: discord.Interaction, member: discord.Member, profiles: list[Record]
@@ -73,8 +96,6 @@ class ProfileCog(commands.Cog, name="Profile"):
             )
             description = []
             for (id_, platform, username) in chunk:
-                if platform == "pc":
-                    username = username.replace("-", "#")
                 description.append(f"{get_platform_emoji(platform)} {username}")
             embed.description = "\n".join(description)
             pages.append(embed)
@@ -85,43 +106,52 @@ class ProfileCog(commands.Cog, name="Profile"):
     async def list(self, interaction: discord.Interaction, member: None | discord.Member = None):
         """List your own or a member's profiles"""
         member = member or interaction.user
-        profiles = await self.get_profiles(interaction, member)
+        profiles = await self.get_profiles(interaction, member.id)
         entries = await self.list_profiles(interaction, member, profiles)
         return await self.bot.paginate(entries, interaction=interaction)
 
+    @can_add_profile()
     @profile.command()
     async def link(self, interaction: discord.Interaction):
         """Link an Overwatch profile"""
         await interaction.response.send_modal(ModalProfileLink())
 
+    @has_profile()
     @profile.command()
     async def update(self, interaction: discord.Interaction):
         """Update an Overwatch profile"""
-        profiles = await self.get_profiles(interaction, interaction.user)
+        profiles = await self.get_profiles(interaction, interaction.user.id)
         await interaction.response.send_modal(ModalProfileUpdate(profiles))
 
+    @has_profile()
     @profile.command()
     async def unlink(self, interaction: discord.Interaction):
         """Unlink an Overwatch profile"""
-        message = "Select a profile to unlink."
-        id_, platform, username = await select_profile(interaction, message)
-        embed = discord.Embed(color=self.bot.color(interaction.user.id))
-        embed.title = "Are you sure you want to unlink the following profile?"
-        embed.add_field(name="Platform", value=platform)
-        embed.add_field(name="Username", value=username)
+        profiles = await self.get_profiles(interaction, interaction.user.id)
+        if len(profiles) == 1:
+            id_, platform, username = profiles[0]
+            embed = discord.Embed(color=self.bot.color(interaction.user.id))
+            embed.title = "Are you sure you want to unlink the following profile?"
+            embed.add_field(name="Platform", value=platform)
+            embed.add_field(name="Username", value=username)
 
-        if await self.bot.prompt(interaction, embed):
-            await interaction.client.pool.execute("DELETE FROM profile WHERE id = $1;", id_)
-            await interaction.followup.send("Profile successfully deleted.", ephemeral=True)
+            if await self.bot.prompt(interaction, embed):
+                await self.bot.pool.execute("DELETE FROM profile WHERE id = $1;", id_)
+                await interaction.response.send_message("Profile successfully unlinked.")
+        else:
+            view = SelectProfilesView(profiles, author_id=interaction.user.id)
+            message = "Select at least a profile to unlink..."
+            view.message = await interaction.response.send_message(message, view=view)
 
     @has_profile()
     @profile.command()
     @app_commands.describe(member="The mention or the ID of a Discord member")
     async def ratings(self, interaction: discord.Interaction, member: None | discord.Member = None):
         """Provides SRs information for a profile"""
+        await interaction.response.defer(thinking=True)
         member = member or interaction.user
         message = "Select a profile to view the skill ratings for."
-        record = await select_profile(interaction, message, member)
+        record = await self.select_profile(interaction, message, member)
         profile = Profile(interaction=interaction, record=record)
         await profile.compute_data()
         if profile.is_private():
@@ -131,7 +161,7 @@ class ProfileCog(commands.Cog, name="Profile"):
             # only update the nickname if the profile matches the one
             # selected for that purpose
             query = "SELECT * FROM nickname WHERE profile_id = $1;"
-            flag = await interaction.client.pool.fetchrow(query, profile.id)
+            flag = await self.bot.pool.fetchrow(query, profile.id)
             if flag and member.id == interaction.user.id:
                 await Nickname(interaction, profile=profile).update()
         await interaction.followup.send(embed=embed)
@@ -141,9 +171,10 @@ class ProfileCog(commands.Cog, name="Profile"):
     @app_commands.describe(member="The mention or the ID of a Discord member")
     async def stats(self, interaction: discord.Interaction, member: None | discord.Member = None):
         """Provides general stats for a profile"""
+        await interaction.response.defer(thinking=True)
         member = member or interaction.user
         message = "Select a profile to view the stats for."
-        _, platform, username = await select_profile(interaction, message, member)
+        _, platform, username = await self.select_profile(interaction, message, member)
         await interaction.client.get_cog("Stats").show_stats_for(
             interaction, "allHeroes", platform, username
         )
@@ -157,9 +188,10 @@ class ProfileCog(commands.Cog, name="Profile"):
         self, interaction: discord.Interaction, hero: str, member: None | discord.Member = None
     ):
         """Provides general hero stats for a profile."""
+        await interaction.response.defer(thinking=True)
         member = member or interaction.user
         message = f"Select a profile to view **{hero}** stats for."
-        _, platform, username = await select_profile(interaction, message, member)
+        _, platform, username = await self.select_profile(interaction, message, member)
         await interaction.client.get_cog("Stats").show_stats_for(
             interaction, hero, platform, username
         )
@@ -169,9 +201,10 @@ class ProfileCog(commands.Cog, name="Profile"):
     @app_commands.describe(member="The mention or the ID of a Discord member")
     async def summary(self, interaction: discord.Interaction, member: None | discord.Member = None):
         """Provides summarized stats for a profile"""
+        await interaction.response.defer(thinking=True)
         member = member or interaction.user
         message = "Select a profile to view the summary for."
-        record = await select_profile(interaction, message, member)
+        record = await self.select_profile(interaction, message, member)
         profile = Profile(interaction=interaction, record=record)
         await profile.compute_data()
         if profile.is_private():
@@ -207,7 +240,7 @@ class ProfileCog(commands.Cog, name="Profile"):
                 )
 
             message = "Select a profile to use for the nickname SRs."
-            record = await select_profile(interaction, message)
+            record = await self.select_profile(interaction, message)
             profile = Profile(interaction=interaction, record=record)
 
             if profile.is_private():
@@ -261,8 +294,6 @@ class ProfileCog(commands.Cog, name="Profile"):
         ax.xaxis.set_major_formatter(matplotlib.dates.DateFormatter("%Y-%m-%d"))
         fig.autofmt_xdate()
 
-        if platform == "pc":
-            username = username.replace("-", "#")
         fig.suptitle(f"{username} - {platform}", fontsize="20")
         pyplot.legend(title="Roles", loc="upper right")
         pyplot.xlabel("Date")
@@ -285,7 +316,7 @@ class ProfileCog(commands.Cog, name="Profile"):
     async def graph(self, interaction: discord.Interaction):
         """Shows SRs performance graph."""
         message = "Select a profile to view the SRs graph for."
-        profile = await select_profile(interaction, message)
+        profile = await self.select_profile(interaction, message)
         file, embed = await self.sr_graph(interaction, profile)
         await interaction.response.send_message(file=file, embed=embed)
